+++ modules/portal/portal-api/src/main/java/com/enonic/xp/portal/websocket/WebSocketManager.java
@@ -6,7 +6,7 @@ public interface WebSocketManager
 
     void sendToGroup( String group, String message );
 
-    long getGroupSize( String group );
+    int getGroupSize( String group );
 
     void addToGroup( String group, String id );
